+++ src/main/java/org/realityforge/gwt/websockets/client/html5/Html5WebSocket.java
@@ -61,6 +61,12 @@ public class Html5WebSocket
   }
 
   @Override
+  public boolean isConnected()
+  {
+    return null != _webSocket;
+  }
+
+  @Override
   public final int getBufferedAmount()
   {
     checkConnected();